fix(remote): percentage panel sizes and draft-safe breakpoint layout
react-resizable-panels v4 treats numeric sizes as pixels, so the right
panel opened at 30-45px wide; sizes are now percentage strings. The
resizable group also renders on mobile (single panel — right panels are
Sheets there), keeping ChatPanel in one tree position across the
responsive breakpoint so composer drafts survive rotation/resize.

Splits websocket helpers out of gateway-integration.test.ts to bring it
back under the 500-LOC rule.
